requirements.txt
# requirements.txt

uuid
threading
asyncio
websocket-client
